Reorganized redux
