+++ src/index.tsx
@@ -6,7 +6,7 @@ import React from 'react';
 import { render } from 'react-dom';
 import { Provider } from 'react-redux';
 import App from './components/app';
-import store from './store';
+import store from './store/store';
 
 const root = document.getElementById('root');
